requirements-minimal.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database and ORM
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
cryptography>=41.0.0,<47.0.0
fastapi-csrf-protect==0.3.2

# Rate limiting and middleware
slowapi==0.1.9

# Image processing (without face-recognition)
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
sentence-transformers==2.2.2

# Cloud storage and external services
cloudinary==1.36.0
aiohttp==3.9.1
requests==2.31.0

# Utilities and data processing
python-slugify==8.0.1
qrcode[pil]==7.4.2
geopy==2.4.1
exifread==3.0.0

# System monitoring
psutil==5.9.6

# Monitoring and metrics
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional dependencies for enhanced features
redis==5.0.1
celery==5.3.4
rq==1.15.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Additional required dependencies
starlette==0.27.0
email-validator==2.1.1
itsdangerous>=2.0.1,<3.0.0